refactor: unify Binance adapter import
