+++ tests/test_live_runner.py
@@ -5,12 +5,14 @@ from types import SimpleNamespace
 import importlib, sys, types
 
 # runner_testnet indirectly imports a module missing in tests; inject a stub
-binance_ws_stub = types.ModuleType("tradingbot.adapters.binance_ws")
-binance_ws_stub.BinanceWSAdapter = object
-sys.modules.setdefault("tradingbot.adapters.binance_ws", binance_ws_stub)
+binance_stub = types.ModuleType("tradingbot.adapters.binance")
+binance_stub.BinanceWSAdapter = object
+sys.modules.setdefault("tradingbot.adapters.binance", binance_stub)
 
 from tradingbot.live import runner_testnet as rt
 from tradingbot.core import normalize, Account
+
+sys.modules.pop("tradingbot.adapters.binance", None)
 
 
 class DummyWS:
